refactor(#3038): move tmux watcher panel decisions
